+++ packages/fabric/editor-core/__tests__/ui/TableFloatingToolbar.tsx
@@ -19,6 +19,10 @@ import {
   tdEmpty,
   tdCursor,
 } from '@atlaskit/editor-test-helpers';
+import {
+  selectRow,
+  selectColumn,
+} from '../../src/editor/plugins/table/actions';
 
 describe('TableFloatingToolbar', () => {
   const event = createEvent('event');
@@ -71,7 +75,7 @@ describe('TableFloatingToolbar', () => {
         doc(p('text'), table(tr(tdCursor, tdEmpty, tdEmpty))),
       );
       plugin.props.handleDOMEvents!.focus(editorView, event);
-      pluginState.selectColumn(0);
+      selectColumn(0)(editorView.state, editorView.dispatch);
       const floatingToolbar = mount(
         <TableFloatingToolbar
           pluginState={pluginState}
@@ -89,7 +93,7 @@ describe('TableFloatingToolbar', () => {
         doc(p('text'), table(tr(tdCursor, tdEmpty, tdEmpty))),
       );
       plugin.props.handleDOMEvents!.focus(editorView, event);
-      pluginState.selectRow(0);
+      selectRow(0)(editorView.state, editorView.dispatch);
       const floatingToolbar = mount(
         <TableFloatingToolbar
           pluginState={pluginState}
@@ -107,7 +111,7 @@ describe('TableFloatingToolbar', () => {
         doc(p('text'), table(tr(tdCursor, tdEmpty, tdEmpty))),
       );
       plugin.props.handleDOMEvents!.focus(editorView, event);
-      pluginState.selectRow(0);
+      selectRow(0)(editorView.state, editorView.dispatch);
       const floatingToolbar = mount(
         <TableFloatingToolbar
           pluginState={pluginState}